Cleaned up lots of build errors with clang++ c++11

+++ desktop/src/segment/lists/omSegmentListsTypes.hpp
@@ -10,7 +10,7 @@ struct SegInfo
     OmSegment* seg;
     OmSegID segID;
     int64_t sizeIncludingChildren;
-    int32_t numChildren;
+    int64_t numChildren;
 };
 
 struct GUIPageOfSegments
